Added start_network to ClientState
